* 'devel' of https://github.com/arangodb/arangodb:
  The Legendary El Cheapo (#8485)
  fix type of hidden startup option `--rocksdb.recycle-log-file-num` (#8517)
  Fix a race in RemoteExecutor (#8516)
  Revert "Feature/ncc1701 (#8440)"
  Revert "fix a compile warning"
  prevent assertion failures when creating smart graphs with invalid options (#8502)
  fix a compile warning
  use abort as SIGABRT has CORE as default action (#8510)
  Make addfollower jobs less aggressive. (#8490)
  Feature/ncc1701 (#8440)

+++ arangod/RocksDBEngine/RocksDBEngine.cpp
@@ -462,7 +462,7 @@ void RocksDBEngine::start() {
   // Maximum number of level-0 files.  We stop writes at this point.
   _options.level0_stop_writes_trigger = static_cast<int>(opts->_level0StopTrigger);
 
-  _options.recycle_log_file_num = static_cast<size_t>(opts->_recycleLogFileNum);
+  _options.recycle_log_file_num = opts->_recycleLogFileNum;
   _options.compaction_readahead_size = static_cast<size_t>(opts->_compactionReadaheadSize);
 
 #ifdef USE_ENTERPRISE
